\name{eigenvalues}
\alias{eigenvalues}
\title{eigenvalues}
\description{This function computes the eigenvalues of a centered and reduced data matrix}
\usage{eigenvalues(X,draw=FALSE)}
\arguments{\item{X}{a matrix of size tm x vm which contains the functionnal images centered and reduced}
  \item{draw}{Logical. Should we plot the eigenvalues}
}
\value{A list containing
\item{eigenvalues}{vector of the eigenvalues}
}
\examples{
# TODO!!
# valpcr <- eigenvalues(Xcr,draw=T)$eigenvalues
}
\keyword{utilities}



